Support Q6_K quantized token embeddings instead of F16
The token embedding tensor is now expected in Q6_K format (from
llama-quantize --token-embedding-type Q6_K). Adds Q6_K block struct
and dequantization routine, updates embedding lookup, LM head dot
product, and weight loading accordingly.

+++ biturbo.h
@@ -155,7 +155,7 @@ typedef struct {
  * ============================================================ */
 
 typedef struct {
-    const uint16_t* token_embedding;  /* [vocab_size, dim] F16 (mmap'd) */
+    const uint8_t* token_embedding;   /* [vocab_size, dim] Q6_K (mmap'd) */
     bt_layer_weights_t* layers;       /* [n_layers]                     */
     float* final_norm;                /* [dim] output RMS norm          */
     /* lm_head: tied to token_embedding (transposed)                    */
